% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTag.R
\name{deTag}
\alias{deTag}
\title{Turn phase tags into columns within PHASE table}
\usage{
deTag(db, textual = T, by.type = T, tickbox = T, include.empty = F)
}
\arguments{
\item{db}{local copy of RATTUS database to work with}

\item{textual}{Logical, default TRUE. Should text columns with concatenated strings of tags be created?}

\item{by.type}{Logical, default TRUE. Should tag strings be split by tag type?}

\item{tickbox}{Logical, default TRUE. Should a grid of logical columns be created to show which tags apply?}

\item{include.empty}{Logical, default FALSE. Should tags that never occur still be given columns?}
}
\value{
A list of tables matching the input, but with new columns in PHASE.
}
\description{
Turn phase tags into columns within PHASE table
}
