% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litStatus.R
\name{litStatus}
\alias{litStatus}
\title{Combine a db object with info from the live progress tracker to summarise the state of play in the literature database module}
\usage{
litStatus(
  db,
  file =
    "https://docs.google.com/spreadsheets/d/1yegDMHO6l2DLcDJtaSfzhMeW8-6kW4kUM9417aLaYh4/edit?gid=613804957#gid=613804957",
  sheet = 2
)
}
\arguments{
\item{db}{local copy of RATTUS database to work with}

\item{file}{URI to current file location}

\item{sheet}{Number of sheet/tab within file}
}
\value{
a list of two data.tables, summarising progress firstly by country and secondly by site.
}
\description{
Combine a db object with info from the live progress tracker to summarise the state of play in the literature database module
}
