% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusDump.R
\name{rattusDump}
\alias{rattusDump}
\title{Dump a database (i.e. list of tables) into a named folder as .csv files}
\usage{
rattusDump(db, name = NULL)
}
\arguments{
\item{db}{List of data.tables as created by one of the rattusPull functions}

\item{name}{Name for folder to write to. If NULL, system date will be used.}
}
\value{
nothing - but writes folder to disk
}
\description{
Dump a database (i.e. list of tables) into a named folder as .csv files
}
