% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusPull.R
\name{rattusPull}
\alias{rattusPull}
\title{Pull in data from the RATTUS Google Sheets-format database.}
\usage{
rattusPull(connection, bestDate = T, bestSpecies = T, prepMetrics = F)
}
\arguments{
\item{connection}{MySQLConnection defining the database connection, or use default for rattus_read connection.}

\item{bestDate}{Should a set of columns be added to SPECIMEN to show the best available dating info drawn from across all tables?}

\item{bestSpecies}{Should a set of columns be added to SPECIMEN to show the most confident species ID available?}

\item{prepMetrics}{Should the fusion status of the relevant end(s) for each metric value be worked out and appended to the METRICS table?}
}
\value{
a list of (currently) 33 data.tables
}
\description{
Pull in data from the RATTUS Google Sheets-format database.
}
