% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusPullFromDisk.R
\name{rattusPullFromDisk}
\alias{rattusPullFromDisk}
\title{Read a previously saved copy of the RATTUS database from disk.
NEED TO ADD IN NEW LOOK-UP TABLES!}
\usage{
rattusPullFromDisk(path)
}
\arguments{
\item{path}{Path to folder containing tables to read.}
}
\value{
a list of (currently) 19 data.tables.
}
\description{
Read a previously saved copy of the RATTUS database from disk.
NEED TO ADD IN NEW LOOK-UP TABLES!
}
