% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusPullGoogle.R
\name{rattusPullGoogle}
\alias{rattusPullGoogle}
\title{Pull in data from the RATTUS Google Sheets-format database.}
\usage{
rattusPullGoogle(
  litMod =
    "https://docs.google.com/spreadsheets/d/1JQIFVbUV3_rRwuOtGL0qbJpR-7EvkFwaBsThC7Zg9Po/edit#gid=0",
  specMod =
    "https://docs.google.com/spreadsheets/d/1TtTrst6ps121h7BNZojJ7R1WzGEeF7jMhoeWzdrH7Mk/edit#gid=0"
)
}
\arguments{
\item{litMod}{URI for the Google sheet containing the site and context tables.}

\item{specMod}{URI  for the Google sheet containing the specimen and sample tables.}
}
\value{
a list of (currently) 23 data.tables (nb. doesn't include documentary source tables)
}
\description{
Pull in data from the RATTUS Google Sheets-format database.
}
