% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusPush.R
\name{rattusPush}
\alias{rattusPush}
\title{Write the contents of a copy of the database out to an existing copy as defined by a MySQLConnection.
This will overwrite the existing contents to use extreme caution. Best used to write to localhost,
after which the results can be checked before restoring the SQL file to the UoY server version.}
\usage{
rattusPush(db, connection, prompt = T, report = T, returnDroppedRows = T)
}
\arguments{
\item{db}{List of tables to write, in format produced by one of the rattusPull functions.}

\item{connection}{MySQLConnection defining the database connection to write to.}

\item{prompt}{Should function prompt user before overwriting data in an existing table?}

\item{report}{Should function provide a running report of tables written and numbers of rows dropped?}

\item{returnDroppedRows}{Should function return a list of tables showing dropped rows?}
}
\value{
if returnDroppedRows is TRUE, a list of one data.table for each table in the databases, showing dropped rows. Otherwise none.
}
\description{
Write the contents of a copy of the database out to an existing copy as defined by a MySQLConnection.
This will overwrite the existing contents to use extreme caution. Best used to write to localhost,
after which the results can be checked before restoring the SQL file to the UoY server version.
}
