% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rattusDecode.R
\name{rattusDecode}
\alias{rattusDecode}
\title{Decode lookup columns in all tables in an R copy of the database. Optionally, also performs merges to pull
in commonly-used key fields from other tables to the PHASE, SPECIMEN, and/or ASSEMBLAGE tables. This option is
intended as a quick-and-dirty shortcut when performing ad hoc queries - it is not recommended in serious analysis
scripts, for which bespoke merges between tables would be mopre appropriate}
\usage{
rattusDecode(db, lookups = T, keyfields = NULL)
}
\arguments{
\item{db}{List of tables to amend, in format produced by one of the rattusPull functions.}

\item{lookups}{Should true lookup fields be decoded? TRUE by default; set to FALSE where already decoded.}

\item{keyfields}{Character vector of tables (from PHASE, SPECIMEN, ASSEMBLAGE) into which key fields should be merged.}
}
\value{
A list of tables matching the input, but with columns merged in the ways defined by lookups and keyfields
}
\description{
Decode lookup columns in all tables in an R copy of the database. Optionally, also performs merges to pull
in commonly-used key fields from other tables to the PHASE, SPECIMEN, and/or ASSEMBLAGE tables. This option is
intended as a quick-and-dirty shortcut when performing ad hoc queries - it is not recommended in serious analysis
scripts, for which bespoke merges between tables would be mopre appropriate
}
